src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.4.1            # via apache-airflow
apache-airflow[crypto,postgres]==1.10.9  # via -r requirements.in
apispec[yaml]==1.3.3      # via flask-appbuilder
argcomplete==1.11.1       # via apache-airflow
attrs==19.3.0             # via apache-airflow, cattrs, jsonschema
babel==2.8.0              # via flask-babel, oslo.i18n
bcrypt==3.1.7             # via flask-bcrypt
cached-property==1.5.1    # via apache-airflow
cattrs==0.9.0             # via -r requirements.in, apache-airflow
certifi==2019.11.28       # via requests
cffi==1.14.0              # via bcrypt, cryptography
chardet==3.0.4            # via requests
click==7.1.1              # via flask, flask-appbuilder
colorama==0.4.3           # via flask-appbuilder
colorlog==4.0.2           # via apache-airflow
configparser==3.5.3       # via apache-airflow
croniter==0.3.31          # via apache-airflow
cryptography==2.8         # via apache-airflow
debtcollector==2.0.0      # via oslo.config, oslo.serialization, oslo.utils, python-keystoneclient
defusedxml==0.6.0         # via python3-openid
dill==0.3.1.1             # via apache-airflow
docutils==0.16            # via python-daemon
dynamic-yaml==1.2.1       # via -r requirements.in
entrypoints==0.3          # via flake8
environs==7.3.0           # via -r requirements.in
flake8==3.7.9             # via -r requirements.in
flask-admin==1.5.4        # via apache-airflow
flask-appbuilder==2.2.4   # via apache-airflow
flask-babel==1.0.0        # via flask-appbuilder
flask-bcrypt==0.7.1       # via -r requirements.in
flask-caching==1.3.3      # via apache-airflow
flask-jwt-extended==3.24.1  # via flask-appbuilder
flask-login==0.4.1        # via apache-airflow, flask-appbuilder
flask-openid==1.2.5       # via flask-appbuilder
flask-sqlalchemy==2.4.1   # via flask-appbuilder
flask-swagger==0.2.13     # via apache-airflow
flask-wtf==0.14.3         # via apache-airflow, flask-appbuilder
flask==1.1.1              # via apache-airflow, flask-admin, flask-appbuilder, flask-babel, flask-bcrypt, flask-caching, flask-jwt-extended, flask-login, flask-openid, flask-sqlalchemy, flask-swagger, flask-wtf
funcsigs==1.0.2           # via apache-airflow
future==0.16.0            # via apache-airflow
graphviz==0.13.2          # via apache-airflow
gunicorn==19.10.0         # via apache-airflow
idna==2.9                 # via requests
iso8601==0.1.12           # via apache-airflow, keystoneauth1, oslo.utils
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.10.3            # via apache-airflow, flask, flask-babel
json-merge-patch==0.2     # via apache-airflow
jsonschema==3.2.0         # via apache-airflow, flask-appbuilder
keystoneauth1==3.18.0     # via python-keystoneclient
lazy-object-proxy==1.4.3  # via apache-airflow
lockfile==0.12.2          # via python-daemon
mako==1.1.2               # via alembic
markdown==2.6.11          # via apache-airflow
markupsafe==1.1.1         # via jinja2, mako
marshmallow-enum==1.5.1   # via flask-appbuilder
marshmallow-sqlalchemy==0.22.3  # via flask-appbuilder
marshmallow==2.19.5       # via environs, flask-appbuilder, marshmallow-enum, marshmallow-sqlalchemy
mccabe==0.6.1             # via flake8
msgpack==1.0.0            # via oslo.serialization
netaddr==0.7.19           # via oslo.config, oslo.utils
netifaces==0.10.9         # via oslo.utils
numpy==1.18.1             # via pandas
os-service-types==1.7.0   # via keystoneauth1
oslo.config==8.0.1        # via python-keystoneclient
oslo.i18n==4.0.0          # via oslo.config, oslo.utils, python-keystoneclient
oslo.serialization==3.1.0  # via python-keystoneclient
oslo.utils==4.1.0         # via oslo.serialization, python-keystoneclient
pandas==0.25.3            # via apache-airflow
pbr==5.4.4                # via debtcollector, keystoneauth1, os-service-types, oslo.i18n, oslo.serialization, oslo.utils, python-keystoneclient, stevedore
pendulum==1.4.4           # via apache-airflow
prison==0.1.2             # via flask-appbuilder
psutil==5.7.0             # via apache-airflow
psycopg2-binary==2.8.4    # via apache-airflow
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via apache-airflow
pyjwt==1.7.1              # via flask-appbuilder, flask-jwt-extended
pyparsing==2.4.6          # via oslo.utils
pyrsistent==0.15.7        # via jsonschema
python-daemon==2.1.2      # via apache-airflow
python-dateutil==2.8.1    # via alembic, apache-airflow, croniter, flask-appbuilder, pandas, pendulum
python-dotenv==0.12.0     # via environs
python-editor==1.0.4      # via alembic
python-keystoneclient==3.22.0  # via -r requirements.in
python-swiftclient==3.9.0  # via -r requirements.in
python3-openid==3.1.0     # via flask-openid
pytz==2019.3              # via babel, flask-babel, oslo.serialization, oslo.utils, pandas, tzlocal
pytzdata==2019.3          # via pendulum
pyyaml==5.3               # via apispec, dynamic-yaml, flask-swagger, oslo.config, oslo.serialization
requests==2.23.0          # via apache-airflow, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
rfc3986==1.3.2            # via oslo.config
setproctitle==1.1.10      # via apache-airflow
six==1.14.0               # via bcrypt, cryptography, debtcollector, flask-jwt-extended, jsonschema, keystoneauth1, oslo.config, oslo.i18n, oslo.serialization, oslo.utils, prison, pyrsistent, python-dateutil, python-keystoneclient, python-swiftclient, sqlalchemy-utils, stevedore, tenacity, thrift
sqlalchemy-jsonfield==0.9.0  # via apache-airflow
sqlalchemy-utils==0.36.1  # via flask-appbuilder
sqlalchemy==1.3.13        # via alembic, apache-airflow, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-jsonfield, sqlalchemy-utils
stevedore==1.32.0         # via keystoneauth1, oslo.config, python-keystoneclient
tabulate==0.8.6           # via apache-airflow
tenacity==4.12.0          # via apache-airflow
termcolor==1.1.0          # via apache-airflow
text-unidecode==1.2       # via apache-airflow
thrift==0.13.0            # via apache-airflow
tzlocal==1.5.1            # via apache-airflow, pendulum
unicodecsv==0.14.1        # via apache-airflow
urllib3==1.25.8           # via requests
werkzeug==0.16.1          # via apache-airflow, flask, flask-caching, flask-jwt-extended
wrapt==1.12.1             # via debtcollector
wtforms==2.2.1            # via flask-admin, flask-wtf
zope.deprecation==4.4.0   # via apache-airflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
